requirements.txt
Flask
pinecone-client
pyvis
python-dotenv